revert: remove broken autofix decompose of grammar.rs
Reverts commits 9d238ff and df6f784 which decomposed grammar.rs into
a module directory with broken imports and verbose filenames. The
decompose naming fix (#975) is now on main, so the next autofix run
will generate correct names.
